# clv_be/user/swagger-docs.json
{"openapi":"3.0.0","paths":{"/auth/users":{"get":{"operationId":"AuthController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDTO"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_getUserLogout","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/user/profile":{"get":{"operationId":"UserController_getUserById","parameters":[],"responses":{"200":{"description":""}},"tags":["user"]}},"/user/activate":{"put":{"operationId":"UserController_activateUserByEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivateUserDto"}}}},"responses":{"202":{"description":""}},"tags":["user"]}},"/user/reset-password":{"put":{"operationId":"UserController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPwDto"}}}},"responses":{"202":{"description":""}},"tags":["user"]}},"/user/list":{"get":{"operationId":"UserController_getAllUsers","parameters":[],"responses":{"200":{"description":""}},"tags":["user"]}},"/user/new-permission":{"post":{"operationId":"UserController_createPermission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PermissionDto"}}}},"responses":{"201":{"description":""}},"tags":["user"]}},"/user/list-permission":{"get":{"operationId":"UserController_getListPermission","parameters":[],"responses":{"200":{"description":""}},"tags":["user"]}},"/user/edit-permission-role":{"put":{"operationId":"UserController_changePermissionRole","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditPermissionDto"}}}},"responses":{"200":{"description":""}},"tags":["user"]}}},"info":{"title":"CLV training","description":"CLV training API description","version":"1.0","contact":{}},"tags":[{"name":"Microservice USER","description":""}],"servers":[],"components":{"schemas":{"RegisterDTO":{"type":"object","properties":{}},"LoginDTO":{"type":"object","properties":{}},"ActivateUserDto":{"type":"object","properties":{}},"ResetPwDto":{"type":"object","properties":{}},"PermissionDto":{"type":"object","properties":{}},"EditPermissionDto":{"type":"object","properties":{}}}}}